Updated mailing list and order request logic. Update db on restart

+++ data-service/src/main/java/com/netcracker/dataservice/service/MailingListService.java
@@ -3,10 +3,12 @@ package com.netcracker.dataservice.service;
 
 import com.netcracker.dataservice.model.MailingList;
 import com.netcracker.dataservice.repositories.MailingListRepository;
+import com.netcracker.dataservice.repositories.OrderRepository;
 import com.netcracker.dataservice.security.JwtService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
 import java.util.Set;
 import java.util.UUID;
 
@@ -16,12 +18,13 @@ public class MailingListService {
 
     private final JwtService jwtService;
     private final MailingListRepository repository;
+    private final OrderRepository orderRepository;
 
     public ResponseEntity<Set<MailingList>> getMailingLists(String token) {
         String jwt = token.substring(7);
         String customerUsername = jwtService.extractUsername(jwt);
         Set<MailingList> mailingLists = repository.findAllByCustomerUsername(customerUsername);
-        mailingLists.forEach(x->x.setQuantityOfClients(x.getClients().size()));
+        mailingLists.forEach(x -> x.setQuantityOfClients(x.getClients().size()));
         return ResponseEntity.ok(mailingLists);
     }
 
@@ -29,9 +32,15 @@ public class MailingListService {
         String jwt = token.substring(7);
         String tokenUsername = jwtService.extractUsername(jwt);
         MailingList listToDelete = repository.findById(UUID.fromString(id)).get();
-        if(listToDelete.getCustomer().getUsername().equals(tokenUsername)){
-            repository.delete(listToDelete);
+        if (listToDelete.getCustomer().getUsername().equals(tokenUsername)) {
+            if (orderRepository.existsByMailingListId(UUID.fromString(id))) {
+                listToDelete.setCustomer(null);
+                repository.save(listToDelete);
+            } else {
+                repository.delete(listToDelete);
+            }
         }
 
     }
+
 }
